Change ComputerPlayer#move to select a center space when available
This optimization increases chances of winning and reduces
computational overhead. This only takes effect for odd sized boards,
since square boards of an even side-length don't have a center space.
If spaces located closer to the center were prioritized in general, it
would be possible to trick the AI into losing by making a winning line
 along an edge. I may try to make this work for even boards.